+++ library/Client/Webhook.php
@@ -23,7 +23,7 @@ class Webhook extends AbstractClient {
                 json_decode($response->getBody(), true, 512, JSON_THROW_ON_ERROR)
             );
         } else {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), (int)esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
 
@@ -39,7 +39,7 @@ class Webhook extends AbstractClient {
             return new \Coinsnap\Result\Webhook($data);
         }
         else {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), (int)esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
 
@@ -76,7 +76,7 @@ class Webhook extends AbstractClient {
             $data = json_decode($response->getBody(), true, 512, JSON_THROW_ON_ERROR);
             return new \Coinsnap\Result\WebhookCreated($data);
         } else {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), (int)esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
 
@@ -125,7 +125,7 @@ class Webhook extends AbstractClient {
             $data = json_decode($response->getBody(), true, 512, JSON_THROW_ON_ERROR);
             return new \Coinsnap\Result\Webhook($data);
         } else {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), (int)esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
 
@@ -148,7 +148,7 @@ class Webhook extends AbstractClient {
         $response = $this->getHttpClient()->request($method, $url, $headers);
 
         if ($response->getStatus() !== 200) {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), (int)esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
 
@@ -173,7 +173,7 @@ class Webhook extends AbstractClient {
             }
             return $r;
         } else {
-            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), esc_html($response->getStatus()), esc_html($response->getBody()));
+            throw $this->getExceptionByStatusCode(esc_html($method), esc_url($url), (int)esc_html($response->getStatus()), esc_html($response->getBody()));
         }
     }
 }
